refactor: Get InputHelper class out of the calibration script so it can be resused!
